cache libffi

+++ pkgs/libffi/fetch.sh
@@ -1,8 +1,11 @@
 #!/bin/sh
 rm -rf "$_DESTDIR" "$_SRCDIR"
-printf "Downloading source...\n"
-curl -L -# -o src.tar.gz https://github.com/libffi/libffi/releases/download/v3.5.2/libffi-3.5.2.tar.gz
+ver='3.5.2'
+if [ ! -f "$_DLCACHE/libffi-$ver.tar.gz" ] ||
+    [ "$(sha256sum "$_DLCACHE/libffi-$ver.tar.gz" | awk '{print $1}')" != "f3a3082a23b37c293a4fcd1053147b371f2ff91fa7ea1b2a52e335676bac82dc" ]; then
+    printf "Downloading source...\n"
+    curl -L -# -o "$_DLCACHE/libffi-$ver.tar.gz" "https://github.com/libffi/libffi/releases/download/v$ver/libffi-$ver.tar.gz" || exit 1
+fi
 printf "Unpacking source...\n"
-tar -C "$_TMP" -xf src.tar.gz
-rm src.tar.gz
+tar -C "$_TMP" -xf "$_DLCACHE/libffi-$ver.tar.gz"
 mv "$_TMP"/libffi-* "$_SRCDIR"
